# spec/openapi/shield-api.v1.yaml
openapi: 3.0.3
info:
  title: Terroir Shield API
  version: 1.0.0
  description: |
    Shield v1 climate-response operations API.
    Includes weather event intake, trigger evaluation, decision workflow,
    case management, basis-risk metrics, and public reporting.
servers:
  - url: http://localhost:8080
    description: Local development server
tags:
  - name: Events
  - name: Triggers
  - name: Decisions
  - name: Cases
  - name: Reports
  - name: Basis Risk

security:
  - bearerAuth: []

paths:
  /v1/events/weather:
    post:
      tags: [Events]
      summary: Ingest a signed weather event packet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WeatherEventPacket'
      responses:
        '202':
          description: Event accepted
          content:
            application/json:
              schema:
                type: object
                required: [accepted, eventId, status]
                properties:
                  accepted: { type: boolean }
                  eventId: { type: string }
                  status: { type: string }
        '400':
          description: Invalid payload

  /v1/programs/{id}/trigger-evaluations:
    post:
      tags: [Triggers]
      summary: Evaluate trigger rules for one enrollment
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [eventId, enrollmentId]
              properties:
                eventId: { type: string }
                enrollmentId: { type: string }
      responses:
        '200':
          description: Trigger and payout preview output
          content:
            application/json:
              schema:
                type: object
                properties:
                  evaluation:
                    $ref: '#/components/schemas/TriggerEvaluation'
                  decision:
                    $ref: '#/components/schemas/PayoutDecision'
                  caseRecord:
                    $ref: '#/components/schemas/CaseRecord'

  /v1/decisions/preview:
    post:
      tags: [Decisions]
      summary: Read current payout preview and optional loss gap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [evaluationId]
              properties:
                evaluationId: { type: string }
                reportedLoss: { type: number, minimum: 0 }
      responses:
        '200':
          description: Preview returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  decision:
                    $ref: '#/components/schemas/PayoutDecision'
                  reportedLoss: { type: number, nullable: true }
                  gap: { type: number, nullable: true }

  /v1/decisions/{id}/approve:
    post:
      tags: [Decisions]
      summary: Approve payout decision and write audit record
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Decision approved
          content:
            application/json:
              schema:
                type: object
                properties:
                  decision:
                    $ref: '#/components/schemas/PayoutDecision'

  /v1/decisions/{id}/export:
    post:
      tags: [Decisions]
      summary: Export approved decision to treasury adapter
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Export payload generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayoutExportBatch'

  /v1/cases:
    get:
      tags: [Cases]
      summary: List operational cases
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [OPEN, UNDER_REVIEW, RESOLVED]
        - in: query
          name: severity
          schema:
            type: string
            enum: [LOW, MEDIUM, HIGH, CRITICAL]
        - in: query
          name: programId
          schema: { type: string }
      responses:
        '200':
          description: Case list
          content:
            application/json:
              schema:
                type: object
                properties:
                  cases:
                    type: array
                    items:
                      $ref: '#/components/schemas/CaseRecord'

  /v1/cases/{id}:
    get:
      tags: [Cases]
      summary: Get full case detail
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Case detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  case:
                    $ref: '#/components/schemas/CaseRecord'
                  event:
                    $ref: '#/components/schemas/WeatherEventPacket'
                  enrollment:
                    $ref: '#/components/schemas/EnrollmentRecord'
                  decision:
                    $ref: '#/components/schemas/PayoutDecision'
                  audit:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLogEntry'

  /v1/reports/public:
    get:
      tags: [Reports]
      summary: Generate aggregated non-PII transparency report
      parameters:
        - in: query
          name: programId
          schema: { type: string }
      responses:
        '200':
          description: Public report
          content:
            application/json:
              schema:
                type: object

  /v1/basis-risk/metrics:
    get:
      tags: [Basis Risk]
      summary: Return basis-risk metrics and history
      parameters:
        - in: query
          name: programId
          schema: { type: string }
      responses:
        '200':
          description: Basis-risk metrics
          content:
            application/json:
              schema:
                type: object
                properties:
                  programId: { type: string }
                  observations: { type: integer }
                  averageExpectedGap: { type: number }
                  latest:
                    $ref: '#/components/schemas/BasisRiskSnapshot'
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/BasisRiskSnapshot'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    ProgramConfig:
      type: object
      required: [programId, name, region, currency, payoutCap, thresholds, weights, effectiveFrom, effectiveTo, ruleVersion]
      properties:
        programId: { type: string }
        name: { type: string }
        region: { type: string }
        currency: { type: string }
        payoutCap: { type: number }
        thresholds:
          type: object
          required: [hailThreshold, frostThreshold, minTemperatureThreshold, droughtThreshold]
          properties:
            hailThreshold: { type: number }
            frostThreshold: { type: number }
            minTemperatureThreshold: { type: number }
            droughtThreshold: { type: number }
        weights:
          type: object
          required: [alpha, beta, gamma]
          properties:
            alpha: { type: number }
            beta: { type: number }
            gamma: { type: number }
        effectiveFrom: { type: string, format: date }
        effectiveTo: { type: string, format: date }
        ruleVersion: { type: string }

    EnrollmentRecord:
      type: object
      required: [enrollmentId, programId, participantRef, batchId, areaHectares, status, trustTier]
      properties:
        enrollmentId: { type: string }
        programId: { type: string }
        participantRef: { type: string }
        batchId: { type: string }
        areaHectares: { type: number }
        status:
          type: string
          enum: [ACTIVE, PENDING, SUSPENDED, CLOSED]
        trustTier:
          type: string
          enum: [SELF_ATTESTED, LAB_ATTESTED, AUTHORITY_ATTESTED]

    WeatherEventPacket:
      type: object
      required: [eventId, programId, hazardType, observedAt, hailIntensity, frostIndicator, minTemperature, droughtIndex, source, signature]
      properties:
        eventId: { type: string }
        programId: { type: string }
        hazardType:
          type: string
          enum: [HAIL, FROST, DROUGHT]
        observedAt: { type: string, format: date-time }
        hailIntensity: { type: number }
        frostIndicator: { type: number }
        minTemperature: { type: number }
        droughtIndex: { type: number }
        source: { type: string }
        signature: { type: string }

    TriggerEvaluation:
      type: object
      required: [evaluationId, eventId, programId, enrollmentId, triggerMet, reasonCodes, requiresManualReview, basisRiskFlag, severityScore, ruleVersion, status, createdAt]
      properties:
        evaluationId: { type: string }
        eventId: { type: string }
        programId: { type: string }
        enrollmentId: { type: string }
        triggerMet: { type: boolean }
        reasonCodes:
          type: array
          items: { type: string }
        requiresManualReview: { type: boolean }
        basisRiskFlag: { type: boolean }
        severityScore: { type: number }
        ruleVersion: { type: string }
        status:
          type: string
          enum: [EVALUATED, REVIEW_REQUIRED]
        createdAt: { type: string, format: date-time }

    PayoutDecision:
      type: object
      required: [decisionId, evaluationId, eventId, programId, enrollmentId, payoutAmount, payoutCap, currency, requiresManualReview, basisRiskFlag, status, ruleVersion, rationale]
      properties:
        decisionId: { type: string }
        evaluationId: { type: string }
        eventId: { type: string }
        programId: { type: string }
        enrollmentId: { type: string }
        payoutAmount: { type: number }
        payoutCap: { type: number }
        currency: { type: string }
        requiresManualReview: { type: boolean }
        basisRiskFlag: { type: boolean }
        status:
          type: string
          enum: [DRAFT, APPROVED, HELD, REJECTED]
        approvedBy: { type: string }
        approvedAt: { type: string, format: date-time }
        ruleVersion: { type: string }
        rationale:
          type: array
          items: { type: string }

    PayoutExportBatch:
      type: object
      required: [exportId, programId, decisionIds, generatedAt, target, signature, status]
      properties:
        exportId: { type: string }
        programId: { type: string }
        decisionIds:
          type: array
          items: { type: string }
        generatedAt: { type: string, format: date-time }
        target: { type: string }
        signature: { type: string }
        status:
          type: string
          enum: [GENERATED, SENT, FAILED]

    BasisRiskSnapshot:
      type: object
      required: [programId, eventId, expectedGap, avgPayout, avgReportedLoss, flaggedCases, generatedAt]
      properties:
        programId: { type: string }
        eventId: { type: string }
        expectedGap: { type: number }
        avgPayout: { type: number }
        avgReportedLoss: { type: number }
        flaggedCases: { type: integer }
        generatedAt: { type: string, format: date-time }

    AuditLogEntry:
      type: object
      required: [auditId, actorId, actorRole, action, entityType, entityId, hash, createdAt]
      properties:
        auditId: { type: string }
        actorId: { type: string }
        actorRole: { type: string }
        action: { type: string }
        entityType: { type: string }
        entityId: { type: string }
        hash: { type: string }
        createdAt: { type: string, format: date-time }

    CaseRecord:
      type: object
      required: [caseId, programId, eventId, enrollmentId, decisionId, severity, status, reason, createdAt]
      properties:
        caseId: { type: string }
        programId: { type: string }
        eventId: { type: string }
        enrollmentId: { type: string }
        decisionId: { type: string }
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        status:
          type: string
          enum: [OPEN, UNDER_REVIEW, RESOLVED]
        reason: { type: string }
        createdAt: { type: string, format: date-time }